fix: disallow keys of JS Object prototype for safety
