+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
-    "target": "es2021",
-    "lib": ["es2021"],
+    "target": "es2022",
+    "lib": ["es2023"],
     "module": "commonjs",
     "sourceMap": true,
     "moduleResolution": "node",
